MakerWare compatibility patch by Steve Weller
Here's a patch to Pleasant3D 2.0.1 that makes it compatible with G-Code created with
MakerWare.

MakerWare omits the 103 and 101 codes so everything ends up one color. The fix tracks the
value of the extruder progress and uses retractions as an indication of travel. It's
backward compatible with the old G-Code as well.