Switched P3DParsedGCode to optimized format for OpenGL VBO drawing
The parsed gcode is now rendered directly into a vertexarray buffer (instead of an array of pane arrays, containing color and Vector3 objects). The vertex array buffer can used directly to create a VBO.
The panes are indexed in a new Array paneIndex. This array contains for each pane the index of its first vertex in the vertex array and the min and max layer height in the pane.
Each vertex in the vertex array buffer contains 4 GLfloat color values (RGBA) and 4 GLfloat coordinate values (XYZW, W=0).